Make clear chat works only on stage and test environments

See merge request vinicius.csantos/no-nsfw-content!8

+++ src/service/clear-chat.service.js
@@ -1,7 +1,12 @@
+require('dotenv').config();
+const ENVIRONMENT = process.env.ENVIRONMENT;
+
 module.exports = (client) => {
-    client.on('message', async message => {
-        if (message.content.startsWith('!clear')) {
-            message.channel.messages.fetch({ limit: 100 }).then(messages => message.channel.bulkDelete(messages, true));
-        }
-    });
+    if (ENVIRONMENT !== 'PROD') {
+        client.on('message', async message => {
+            if (message.content.startsWith('!clear')) {
+                message.channel.messages.fetch({ limit: 100 }).then(messages => message.channel.bulkDelete(messages, true));
+            }
+        });
+    }
 }
